feat: add navigator for stack navigation and create and sample screen for stack navigation
Navigator is a component provided by the createNativeStackNavigator function native-stack. It encloses all the screens that are to be navigated. Sreen component describes a single named component that will be rendered if a particular route is mounted. A Screen component is also provided by the createNativeStackNavigator function and must have a name property that will be used to navigate to that stack screen.